# manager-zedaapi/public/openapi-content-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zé da API - Content Management API",
    "description": "API completa para gerenciamento externo de conteudo da plataforma Zé da API.\n\nTodos os endpoints requerem autenticacao via header `Authorization: Bearer <CONTENT_API_KEY>`.\n\nA chave e definida na variavel de ambiente `CONTENT_API_KEY` do servidor.",
    "version": "1.0.0",
    "contact": {
      "name": "Zé da API",
      "url": "https://zedaapi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento local"
    },
    {
      "url": "https://app.zedaapi.com",
      "description": "Producao"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CONTENT_API_KEY definida no .env do servidor. Minimo 32 caracteres."
      }
    },
    "schemas": {
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5abc123def456"
          },
          "title": {
            "type": "string",
            "example": "Como integrar WhatsApp API em 5 minutos"
          },
          "slug": {
            "type": "string",
            "example": "como-integrar-whatsapp-api-em-5-minutos"
          },
          "content": {
            "type": "string",
            "example": "<h2>Introducao</h2><p>Neste artigo voce vai aprender...</p>"
          },
          "excerpt": {
            "type": "string",
            "nullable": true,
            "example": "Aprenda a integrar a WhatsApp API rapidamente com a Zé da API."
          },
          "coverImageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://s3.example.com/blog/2026/02/cover.webp"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "example": "draft"
          },
          "seoTitle": {
            "type": "string",
            "nullable": true,
            "example": "Integrar WhatsApp API - Guia Completo 2026"
          },
          "seoDescription": {
            "type": "string",
            "nullable": true,
            "example": "Guia passo a passo para integrar a WhatsApp Business API usando a Zé da API."
          },
          "seoKeywords": {
            "type": "string",
            "nullable": true,
            "example": "whatsapp api, integracao, tutorial, webhook"
          },
          "readingTimeMin": {
            "type": "integer",
            "example": 5
          },
          "viewCount": {
            "type": "integer",
            "example": 142
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/BlogCategory"
          },
          "author": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string",
                "example": "Admin"
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "tagId": {
                  "type": "string"
                },
                "postId": {
                  "type": "string"
                },
                "tag": {
                  "$ref": "#/components/schemas/BlogTag"
                }
              }
            }
          }
        }
      },
      "BlogCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5cat001"
          },
          "name": {
            "type": "string",
            "example": "Tutoriais"
          },
          "slug": {
            "type": "string",
            "example": "tutoriais"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Guias passo a passo para usar a plataforma"
          },
          "sortOrder": {
            "type": "integer",
            "example": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "integer",
                "example": 12
              }
            }
          }
        }
      },
      "BlogTag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "example": "cm5tag001"
          },
          "name": {
            "type": "string",
            "example": "WhatsApp"
          },
          "slug": {
            "type": "string",
            "example": "whatsapp"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "posts": {
                "type": "integer",
                "example": 8
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "API key ausente ou invalida",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "missing": {
                "summary": "Header ausente",
                "value": {
                  "error": "Missing Authorization header"
                }
              },
              "invalid": {
                "summary": "Chave invalida",
                "value": {
                  "error": "Invalid API key"
                }
              }
            }
          }
        }
      },
      "NotConfigured": {
        "description": "CONTENT_API_KEY nao configurada no servidor",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Content API is not configured"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflito (slug duplicado)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "A post with this slug already exists"
            }
          }
        }
      }
    }
  },
...
  "paths": {
    "/api/admin/content/blog/posts": {
      "get": {
        "tags": [
          "Blog Posts"
        ],
        "summary": "Listar posts do blog",
        "description": "Retorna lista paginada de posts com filtros opcionais por busca, categoria e status.",
        "operationId": "listBlogPosts",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Numero da pagina (comeca em 1)",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            },
            "example": 1
          },
          {
            "name": "search",
            "in": "query",
            "description": "Busca por titulo (case-insensitive)",
            "schema": {
              "type": "string"
            },
            "example": "whatsapp"
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "Filtrar por ID da categoria",
            "schema": {
              "type": "string"
            },
            "example": "cm5cat001"
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "archived"
              ]
            },
            "example": "published"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BlogPost"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 47
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "pageSize": {
                      "type": "integer",
                      "example": 20
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 3
                    }
                  }
                },
                "example": {
                  "items": [
                    {
                      "id": "cm5abc123def456",
                      "title": "Como integrar WhatsApp API em 5 minutos",
                      "slug": "como-integrar-whatsapp-api-em-5-minutos",
                      "content": "<h2>Introducao</h2><p>Neste artigo...</p>",
                      "excerpt": "Aprenda a integrar rapidamente.",
                      "coverImageUrl": "https://s3.example.com/blog/2026/02/cover.webp",
                      "status": "published",
                      "readingTimeMin": 5,
                      "viewCount": 142,
                      "publishedAt": "2026-02-15T10:00:00.000Z",
                      "createdAt": "2026-02-15T09:30:00.000Z",
                      "updatedAt": "2026-02-15T10:00:00.000Z",
                      "category": {
                        "id": "cm5cat001",
                        "name": "Tutoriais",
                        "slug": "tutoriais"
                      },
                      "author": {
                        "id": "cm5usr001",
                        "name": "Admin"
                      },
                      "tags": [
                        {
                          "tagId": "cm5tag001",
                          "postId": "cm5abc123def456",
                          "tag": {
                            "id": "cm5tag001",
                            "name": "WhatsApp",
                            "slug": "whatsapp"
                          }
                        }
                      ]
                    }
                  ],
                  "total": 47,
                  "page": 1,
                  "pageSize": 20,
                  "totalPages": 3
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/NotConfigured"
          }
        }
      },
      "post": {
        "tags": [
          "Blog Posts"
        ],
        "summary": "Criar post no blog",
        "description": "Cria um novo post. O slug e gerado automaticamente a partir do titulo. O autor e atribuido ao primeiro admin do sistema. Tags sao vinculadas por ID.",
        "operationId": "createBlogPost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Titulo do post",
                    "example": "Como usar webhooks na Zé da API"
                  },
                  "content": {
                    "type": "string",
                    "description": "Conteudo HTML do post",
                    "example": "<h2>O que sao webhooks?</h2><p>Webhooks permitem receber notificacoes em tempo real...</p><h2>Configurando</h2><p>Acesse o painel e vá em Configuracoes > Webhooks...</p>"
                  },
                  "excerpt": {
                    "type": "string",
                    "description": "Resumo curto para listagens",
                    "example": "Aprenda a configurar webhooks para receber mensagens em tempo real."
                  },
                  "coverImageUrl": {
                    "type": "string",
                    "description": "URL da imagem de capa (use /upload primeiro)",
                    "example": "https://s3.example.com/blog/2026/02/uuid-cover.webp"
                  },
                  "categoryId": {
                    "type": "string",
                    "description": "ID da categoria",
                    "example": "cm5cat001"
                  },
                  "tagIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "IDs das tags a vincular",
                    "example": [
                      "cm5tag001",
                      "cm5tag002"
                    ]
                  },
                  "seoTitle": {
                    "type": "string",
                    "description": "Titulo SEO (meta title)",
                    "example": "Webhooks WhatsApp API - Guia Completo | Zé da API"
                  },
                  "seoDescription": {
                    "type": "string",
                    "description": "Descricao SEO (meta description)",
                    "example": "Guia completo sobre como configurar e usar webhooks na Zé da API para receber mensagens WhatsApp em tempo real."
                  },
                  "seoKeywords": {
                    "type": "string",
                    "description": "Palavras-chave SEO separadas por virgula",
                    "example": "webhook, whatsapp api, notificacao, tempo real"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published"
                    ],
                    "description": "Status inicial (default: draft)",
                    "example": "draft"
                  }
                }
              },
              "examples": {
                "rascunho": {
                  "summary": "Criar como rascunho",
                  "value": {
                    "title": "Como usar webhooks na Zé da API",
                    "content": "<h2>O que sao webhooks?</h2><p>Webhooks permitem receber notificacoes em tempo real quando uma mensagem chega no seu numero WhatsApp.</p><h2>Configurando</h2><p>1. Acesse o painel</p><p>2. Va em Configuracoes > Webhooks</p><p>3. Insira a URL do seu servidor</p>",
                    "excerpt": "Aprenda a configurar webhooks para receber mensagens em tempo real.",
                    "categoryId": "cm5cat001",
                    "tagIds": [
                      "cm5tag001",
                      "cm5tag002"
                    ],
                    "seoTitle": "Webhooks WhatsApp API - Guia Completo | Zé da API",
                    "seoDescription": "Guia completo sobre como configurar webhooks na Zé da API.",
                    "seoKeywords": "webhook, whatsapp api, notificacao"
                  }
                },
                "publicar-direto": {
                  "summary": "Criar e publicar direto",
                  "value": {
                    "title": "Novidades da Zé da API - Fevereiro 2026",
                    "content": "<h2>O que ha de novo</h2><p>Neste mes lancamos diversas melhorias...</p>",
                    "excerpt": "Confira as novidades de fevereiro 2026.",
                    "coverImageUrl": "https://s3.example.com/blog/2026/02/cover-novidades.webp",
                    "categoryId": "cm5cat002",
                    "status": "published",
                    "seoTitle": "Novidades Fevereiro 2026 | Zé da API",
                    "seoDescription": "Veja tudo que lancamos em fevereiro 2026 na Zé da API."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                },
                "example": {
                  "id": "cm5newpost001",
                  "title": "Como usar webhooks na Zé da API",
                  "slug": "como-usar-webhooks-na-ze-da-api",
                  "content": "<h2>O que sao webhooks?</h2><p>Webhooks permitem...</p>",
                  "excerpt": "Aprenda a configurar webhooks para receber mensagens em tempo real.",
                  "coverImageUrl": null,
                  "status": "draft",
                  "readingTimeMin": 3,
                  "viewCount": 0,
                  "publishedAt": null,
                  "createdAt": "2026-02-21T12:00:00.000Z",
                  "updatedAt": "2026-02-21T12:00:00.000Z",
                  "category": {
                    "id": "cm5cat001",
                    "name": "Tutoriais",
                    "slug": "tutoriais"
                  },
                  "tags": [
                    {
                      "tagId": "cm5tag001",
                      "postId": "cm5newpost001",
                      "tag": {
                        "id": "cm5tag001",
                        "name": "WhatsApp",
                        "slug": "whatsapp"
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Campos obrigatorios faltando",
            "content": {
              "application/json": {
                "example": {
                  "error": "title and content are required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "Slug duplicado",
            "content": {
              "application/json": {
                "example": {
                  "error": "A post with this slug already exists"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/content/blog/tags": {
      "get": {
        "tags": [
          "Blog Tags"
        ],
        "summary": "Listar tags do blog",
        "description": "Retorna todas as tags ordenadas por nome, com contagem de posts.",
        "operationId": "listBlogTags",
        "responses": {
          "200": {
            "description": "Lista de tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogTag"
                  }
                },
                "example": [
                  {
                    "id": "cm5tag001",
                    "name": "WhatsApp",
                    "slug": "whatsapp",
                    "_count": {
                      "posts": 15
                    }
                  },
                  {
                    "id": "cm5tag002",
                    "name": "Webhook",
                    "slug": "webhook",
                    "_count": {
                      "posts": 8
                    }
                  },
                  {
                    "id": "cm5tag003",
                    "name": "API",
                    "slug": "api",
                    "_count": {
                      "posts": 20
                    }
                  },
                  {
                    "id": "cm5tag004",
                    "name": "Tutorial",
                    "slug": "tutorial",
                    "_count": {
                      "posts": 10
                    }
                  }
                ]
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Blog Tags"
        ],
        "summary": "Criar tag do blog",
        "description": "Cria uma nova tag. O slug e gerado automaticamente a partir do nome.",
        "operationId": "createBlogTag",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Automacao"
                  }
                }
              },
              "example": {
                "name": "Automacao"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogTag"
                },
                "example": {
                  "id": "cm5tag005",
                  "name": "Automacao",
                  "slug": "automacao",
                  "_count": {
                    "posts": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Nome obrigatorio",
            "content": {
              "application/json": {
                "example": {
                  "error": "name is required"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
  }
}